requirements.txt
asgiref==3.7.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
cx-Oracle==8.3.0
distlib==0.3.6
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
Faker==18.10.1
filelock==3.12.2
Flask==2.3.2
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.0
oracledb==1.3.2
Pillow==10.0.0
platformdirs==3.8.0
pycparser==2.21
PyMySQL==1.0.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.16
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
Unidecode==1.3.7
urllib3==2.0.3
virtualenv==20.23.1
Werkzeug==2.3.4
